refactor: update ndk handling
