+++ package.json
@@ -55,6 +55,8 @@
   "devDependencies": {
     "@capacitor/assets": "^3.0.5",
     "@capacitor/cli": "^6.2.0",
+    "@noble/curves": "^1.9.2",
+    "@noble/secp256k1": "^2.3.0",
     "@quasar/app-vite": "^1.9.3",
     "@types/node": "^20.12.11",
     "@types/underscore": "^1.11.15",
